feat: 🎸 Renamed File [AVATAR UTIL]
